+++ pkg/clientset/typed/configuration/v1/configuration_client.go
@@ -30,7 +30,6 @@ type ConfigurationV1Interface interface {
 	RESTClient() rest.Interface
 	KongClusterPluginsGetter
 	KongConsumersGetter
-	KongIngressesGetter
 	KongPluginsGetter
 }
 
@@ -45,10 +44,6 @@ func (c *ConfigurationV1Client) KongClusterPlugins() KongClusterPluginInterface 
 
 func (c *ConfigurationV1Client) KongConsumers(namespace string) KongConsumerInterface {
 	return newKongConsumers(c, namespace)
-}
-
-func (c *ConfigurationV1Client) KongIngresses(namespace string) KongIngressInterface {
-	return newKongIngresses(c, namespace)
 }
 
 func (c *ConfigurationV1Client) KongPlugins(namespace string) KongPluginInterface {