+++ pkg/clientset/typed/configuration/v1beta1/configuration_client.go
@@ -30,8 +30,6 @@ type ConfigurationV1beta1Interface interface {
 	RESTClient() rest.Interface
 	KongConsumerGroupsGetter
 	KongUpstreamPoliciesGetter
-	TCPIngressesGetter
-	UDPIngressesGetter
 }
 
 // ConfigurationV1beta1Client is used to interact with features provided by the configuration.konghq.com group.
@@ -45,14 +43,6 @@ func (c *ConfigurationV1beta1Client) KongConsumerGroups(namespace string) KongCo
 
 func (c *ConfigurationV1beta1Client) KongUpstreamPolicies(namespace string) KongUpstreamPolicyInterface {
 	return newKongUpstreamPolicies(c, namespace)
-}
-
-func (c *ConfigurationV1beta1Client) TCPIngresses(namespace string) TCPIngressInterface {
-	return newTCPIngresses(c, namespace)
-}
-
-func (c *ConfigurationV1beta1Client) UDPIngresses(namespace string) UDPIngressInterface {
-	return newUDPIngresses(c, namespace)
 }
 
 // NewForConfig creates a new ConfigurationV1beta1Client for the given config.